projects: max31827-evkit: migrate to CMake
Replace the Make-based build with CMake for all supported
host platforms (linux-userspace/usb2pmb2, win, mac):
  - Remove Makefile, src.mk, and src/platform/linux/
  - Add src/platform/linux-userspace/ for the usb2pmb2 board
    (ftd2xx I2C, conditional on CONFIG_FTD2XX)
  - Update src/platform/win/ to use ftd2xx unconditionally,
    matching the mac platform
  - Add boards/basic/ and boards/iio_example/ board conf files
    for usb2pmb2, win-usb2pmb2 and mac-usb2pmb2
  - Set builds.json to {} (CMake auto-discovers all combinations)

Tested on linux-userspace/usb2pmb2: basic reads 25.5 °C;
iio_example exposes full IIO context via iio_info.
