--- projects/max31827-evkit/Makefile
@@ -1,11 +1,0 @@
-# Select the example you want to enable by choosing y for enabling and n for disabling
-EXAMPLE ?= basic
-LIBRARIES += ftd2xx
-
-include ../../tools/scripts/generic_variables.mk
-
-include src.mk
-
-include ../../tools/scripts/examples.mk
-
-include ../../tools/scripts/generic.mk
