--- projects/max31827-evkit/builds.json
@@ -1,18 +1,0 @@
-{
-    "linux": {
-      "basic_example": {
-        "flags" : "EXAMPLE=basic"
-      },
-      "iio_example": {
-        "flags" : "EXAMPLE=iio_example"
-      }
-    },
-    "mac": {
-      "basic_example": {
-        "flags" : "EXAMPLE=basic"
-      },
-      "iio_example": {
-        "flags" : "EXAMPLE=iio_example"
-      }
-    }
-  }
