+++ projects/max31827-evkit/src/platform/linux-userspace/parameters.c
@@ -1,6 +1,6 @@
 /*******************************************************************************
- *   @file   parameters.h
- *   @brief  Definition of linux platform data used by max31827 project.
+ *   @file   parameters.c
+ *   @brief  Definition of linux-userspace platform parameters for max31827-evkit.
  *   @author Radu Sabau (radu.sabau@analog.com)
  *******************************************************************************
  * Copyright 2025(c) Analog Devices, Inc.
@@ -19,7 +19,7 @@
  *    contributors may be used to endorse or promote products derived from this
  *    software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
+ * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. "AS IS" AND ANY EXPRESS OR
  * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
  * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
  * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
@@ -30,37 +30,13 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  ******************************************************************************/
-#ifndef __PARAMETERS_H__
-#define __PARAMETERS_H__
+#include "parameters.h"
 
-#ifdef FTD2XX
-#include "ftd2xx_uart.h"
-#include "ftd2xx_i2c.h"
-#else
-#include "linux_i2c.h"
-#include "linux_uart.h"
+#ifdef CONFIG_FTD2XX
+struct ftd2xx_i2c_init max31827_i2c_extra = {
+	.channel_config = false,
+	.extra_pins = true,
+	.extra_pins_dir = 0xFF,
+	.extra_pins_val = 0xBE
+};
 #endif
-
-#define UART_IRQ_ID	0
-#define UART_DEVICE_ID	0
-#define UART_BAUDRATE	0
-#define UART_EXTRA      NULL
-
-#define I2C_DEVICE_ID   0
-
-#ifdef FTD2XX
-#define I2C_OPS         &ftd2xx_i2c_ops
-#define UART_OPS        &ftd2xx_uart_ops
-#else
-#define I2C_OPS		&linux_i2c_ops
-#define UART_OPS	&linux_uart_ops
-#endif
-#define I2C_EXTRA       &max31827_i2c_extra
-
-#ifdef FTD2XX
-extern struct ftd2xx_i2c_init max31827_i2c_extra;
-#else
-extern struct linux_i2c_init_param max31827_i2c_extra;
-#endif
-
-#endif /* __PARAMETERS_H__ */
