+++ projects/max31827-evkit/src/platform/win/parameters.c
@@ -1,6 +1,6 @@
 /*******************************************************************************
  *   @file   parameters.c
- *   @brief  Definition of linux platform data used by max31827 project.
+ *   @brief  Definition of win platform parameters for max31827-evkit.
  *   @author Radu Sabau (radu.sabau@analog.com)
  *******************************************************************************
  * Copyright 2025(c) Analog Devices, Inc.
@@ -19,7 +19,7 @@
  *    contributors may be used to endorse or promote products derived from this
  *    software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
+ * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. "AS IS" AND ANY EXPRESS OR
  * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
  * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
  * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
@@ -32,11 +32,9 @@
  ******************************************************************************/
 #include "parameters.h"
 
-#ifdef FTD2XX
 struct ftd2xx_i2c_init max31827_i2c_extra = {
 	.channel_config = false,
 	.extra_pins = true,
 	.extra_pins_dir = 0xFF,
 	.extra_pins_val = 0xBE
 };
-#endif
