+++ projects/max31827-evkit/src/platform/win/parameters.h
@@ -1,6 +1,6 @@
 /*******************************************************************************
  *   @file   parameters.h
- *   @brief  Definition of linux platform data used by max31827 project.
+ *   @brief  Definitions of win platform parameters for max31827-evkit.
  *   @author Radu Sabau (radu.sabau@analog.com)
  *******************************************************************************
  * Copyright 2025(c) Analog Devices, Inc.
@@ -19,7 +19,7 @@
  *    contributors may be used to endorse or promote products derived from this
  *    software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
+ * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. "AS IS" AND ANY EXPRESS OR
  * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
  * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
  * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
@@ -33,22 +33,20 @@
 #ifndef __PARAMETERS_H__
 #define __PARAMETERS_H__
 
-#ifdef FTD2XX
 #include "ftd2xx_uart.h"
 #include "ftd2xx_i2c.h"
 
 #define UART_IRQ_ID	0
 #define UART_DEVICE_ID	0
 #define UART_BAUDRATE	0
-#define UART_EXTRA      NULL
+#define UART_EXTRA	NULL
 
-#define I2C_DEVICE_ID   0
+#define I2C_DEVICE_ID	0
 
-#define I2C_OPS         &ftd2xx_i2c_ops
-#define UART_OPS        &ftd2xx_uart_ops
+#define I2C_OPS		&ftd2xx_i2c_ops
+#define UART_OPS	&ftd2xx_uart_ops
+#define I2C_EXTRA	&max31827_i2c_extra
 
-#define I2C_EXTRA       &max31827_i2c_extra
 extern struct ftd2xx_i2c_init max31827_i2c_extra;
-#endif
 
 #endif /* __PARAMETERS_H__ */
